// lib/generated/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  CLIENTE: 'CLIENTE',
  BARBEIRO: 'BARBEIRO',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserPlan = {
  START: 'START',
  PREMIUM: 'PREMIUM'
} as const

export type UserPlan = (typeof UserPlan)[keyof typeof UserPlan]